requirements.txt
fastapi>=0.115.0,<1
motor>=3.4.0,<4.0.0
python-dotenv>=1.0.1,<2
python-multipart>=0.0.1,<1
pyjwt>=2.8.0,<3
requests>=2.31.0,<3
uvicorn>=0.29.0
